+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+ * Synfig importer improvements
 
 ## 0.6.5 - 2020-07-10
  * Text rendering improvements