+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
  * Synfig importer improvements
+ * SVG exporter improvements
 
 ## 0.6.5 - 2020-07-10
  * Text rendering improvements